Move gcf functions out of bigtable_automation
