+++ gcf/common.go
