+++ gcf/custom_dc/deploy.sh
