+++ gcf/custom_dc/deploy_local.sh
