+++ gcf/custom_dc/test.sh
