+++ gcf/local/deploy.sh
