+++ gcf/local/test.sh
