+++ gcf/proto/dc_manifest.proto
